Allow inplacing buffer when other users are inconsequential
Summary:
We can inplace a buffer if all of the users of said buffer are
"inconsequential", defined as having been removed, being completed, being part
of the current fused node, or are part of the ancestors set. In particular, this
allows LayerNorm to inplace.

Fixes:
https://github.com/pytorch/pytorch/issues/132826

Test Plan:
New unit test of matmul followed by LayerNorm, make sure there's an inplaced
buffer.